+++ crates/jacquard-common/src/lib.rs
@@ -221,6 +221,8 @@ pub mod types;
 // XRPC protocol types and traits
 pub mod xrpc;
 
+pub use types::value::*;
+
 /// Authorization token types for XRPC requests.
 #[derive(Debug, Clone)]
 pub enum AuthorizationToken<'s> {